Fix nightly updating to stable

+++ cmd/installer/installer.go
@@ -8,12 +8,15 @@ import (
 	"path"
 	"path/filepath"
 	"runtime"
+	"strings"
 
 	"github.com/renproject/swapperd/driver/keystore"
 	"github.com/renproject/swapperd/driver/service"
 	"github.com/renproject/swapperd/driver/updater"
 	"github.com/tyler-smith/go-bip39"
 )
+
+var version = "undefined"
 
 const reset = "\033[m"
 const cyan = "\033[36m"
@@ -50,6 +53,9 @@ func main() {
 }
 
 func updateSwapperd(binDir string) error {
+	if strings.Contains(version, "nightly") && updater.DisableNightlyAutoUpdate {
+		return nil
+	}
 	service.Stop("swapperd-updater")
 	service.Stop("swapperd")
 	service.Delete("swapperd-updater")
